backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
google-generativeai==0.3.1
google-api-python-client==2.108.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.31.0
beautifulsoup4==4.12.2
httpx==0.25.1
python-multipart==0.0.6
lxml==4.9.3